Causing another earthquake to add static lib to the repo & build chain
