+++ Game_Of_Life.sln
@@ -1,9 +1,14 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.6.33712.159
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.33529.622
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Game_Of_Life", "Game_Of_Life.vcxproj", "{1554793B-85FA-4C30-87B8-A091871EF733}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "libFunni", "libFunni\libFunni.vcxproj", "{C5B941B3-447F-471A-A562-D06A1CBB5599}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Game_Of_Life", "Game_Of_Life\Game_Of_Life.vcxproj", "{1554793B-85FA-4C30-87B8-A091871EF733}"
+	ProjectSection(ProjectDependencies) = postProject
+		{C5B941B3-447F-471A-A562-D06A1CBB5599} = {C5B941B3-447F-471A-A562-D06A1CBB5599}
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,6 +18,14 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C5B941B3-447F-471A-A562-D06A1CBB5599}.Debug|x64.ActiveCfg = Debug|x64
+		{C5B941B3-447F-471A-A562-D06A1CBB5599}.Debug|x64.Build.0 = Debug|x64
+		{C5B941B3-447F-471A-A562-D06A1CBB5599}.Debug|x86.ActiveCfg = Debug|Win32
+		{C5B941B3-447F-471A-A562-D06A1CBB5599}.Debug|x86.Build.0 = Debug|Win32
+		{C5B941B3-447F-471A-A562-D06A1CBB5599}.Release|x64.ActiveCfg = Release|x64
+		{C5B941B3-447F-471A-A562-D06A1CBB5599}.Release|x64.Build.0 = Release|x64
+		{C5B941B3-447F-471A-A562-D06A1CBB5599}.Release|x86.ActiveCfg = Release|Win32
+		{C5B941B3-447F-471A-A562-D06A1CBB5599}.Release|x86.Build.0 = Release|Win32
 		{1554793B-85FA-4C30-87B8-A091871EF733}.Debug|x64.ActiveCfg = Debug|x64
 		{1554793B-85FA-4C30-87B8-A091871EF733}.Debug|x64.Build.0 = Debug|x64
 		{1554793B-85FA-4C30-87B8-A091871EF733}.Debug|x86.ActiveCfg = Debug|Win32
