+++ libFunni/libFunni.cpp
@@ -1,0 +1,10 @@
+// libFunni.cpp : Defines the functions for the static library.
+//
+
+#include "pch.h"
+#include "framework.h"
+
+// TODO: This is an example of a library function
+void fnlibFunni()
+{
+}
